res/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=3.5.0=py37haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py37h2bbff1b_0
attrs=22.1.0=py37haa95532_0
babel=2.11.0=py37haa95532_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py37haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blosc=1.21.0=h19a0ad4_1
bottleneck=1.3.5=py37h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py37h2bbff1b_1003
bzip2=1.0.8=he774522_0
ca-certificates=2023.05.30=haa95532_0
category_encoders=2.5.0=py37hd4e2768_1
certifi=2022.12.7=py37haa95532_0
cffi=1.15.1=py37h2bbff1b_3
cfitsio=3.470=h2bbff1b_7
charls=2.2.0=h6c2663c_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.5=py37haa95532_0
console_shortcut=0.1.1=4
cryptography=38.0.1=py37h21b164f_0
curl=7.86.0=h2bbff1b_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.12.0=py37h2bbff1b_0
dask-core=2021.10.0=pyhd3eb1b0_0
debugpy=1.5.1=py37hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
double-conversion=3.1.5=ha925a31_1
eigen=3.3.7=h59b6b97_1
entrypoints=0.4=py37haa95532_0
expat=2.4.9=h6c2663c_0
fftw=3.3.9=h2bbff1b_1
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
fsspec=2022.11.0=py37haa95532_0
giflib=5.2.1=h62dcd97_0
gl2ps=1.4.2=hce2df77_1
glew=2.1.0=hd77b12b_3
glib=2.69.1=h5dc1a3c_2
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
hdf4=4.2.13=h712560f_2
hdf5=1.10.6=h1756f20_1
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py37haa95532_0
imagecodecs=2021.8.26=py37hc0a7faf_1
imageio=2.19.3=py37haa95532_0
importlib-metadata=4.11.3=py37haa95532_0
importlib_metadata=4.11.3=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.15.2=py37haa95532_0
ipython=7.31.1=py37haa95532_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py37haa95532_1
jinja2=3.1.2=py37haa95532_0
joblib=1.1.1=py37haa95532_0
jpeg=9e=h2bbff1b_0
json5=0.9.6=pyhd3eb1b0_0
jsoncpp=1.9.4=h59b6b97_2
jsonschema=4.16.0=py37haa95532_0
jupyter=1.0.0=py37haa95532_8
jupyter_client=7.4.8=py37haa95532_0
jupyter_console=6.4.4=py37haa95532_0
jupyter_core=4.11.2=py37haa95532_0
jupyter_server=1.23.4=py37haa95532_0
jupyterlab=3.5.2=py37haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.16.5=py37haa95532_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.2=py37hd77b12b_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libcurl=7.86.0=h86230a5_0
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
libnetcdf=4.8.1=h6685c40_1
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.20.1=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libssh2=1.10.0=hcd4344a_0
libtheora=1.1.1=h8cc25b3_3
libtiff=4.4.0=h8a3f274_2
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
libzip=1.8.0=h49b8836_0
libzopfli=1.0.3=ha925a31_0
locket=1.0.0=py37haa95532_0
loguru=0.5.3=py37haa95532_3
lxml=4.9.1=py37h1985fb9_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.1=py37h2bbff1b_0
matplotlib=3.5.2=py37haa95532_0
matplotlib-base=3.5.2=py37hd77b12b_0
matplotlib-inline=0.1.6=py37haa95532_0
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.1=py37h277e83a_0
mkl_random=1.2.2=py37hf11a4ad_0
munkres=1.1.4=py_0
nbclassic=0.4.8=py37haa95532_0
nbclient=0.5.13=py37haa95532_0
nbconvert=6.5.4=py37haa95532_0
nbformat=5.7.0=py37haa95532_0
nest-asyncio=1.5.5=py37haa95532_0
networkx=2.6.3=pyhd3eb1b0_0
notebook=6.5.2=py37haa95532_0
notebook-shim=0.2.2=py37haa95532_0
numexpr=2.8.4=py37h5b0cc5e_0
numpy=1.21.5=py37h7a0a035_3
numpy-base=1.21.5=py37hca35cd5_3
opencv=4.6.0=py37h104de81_2
openjpeg=2.4.0=h4fc8c34_0
openssl=1.1.1u=h2bbff1b_0
packaging=22.0=py37haa95532_0
pandas=1.3.5=py37h6214cd6_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_1
patsy=0.5.3=py37haa95532_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.3.0=py37hdc2b20a_1
pip=22.3.1=py37haa95532_0
pkgutil-resolve-name=1.3.10=py37haa95532_0
ply=3.11=py37_0
proj=6.2.1=h3758d61_0
prometheus_client=0.14.1=py37haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
psutil=5.9.0=py37h2bbff1b_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py37haa95532_0
pyqt=5.15.7=py37hd77b12b_0
pyqt5-sip=12.11.0=py37hd77b12b_0
pyrsistent=0.18.0=py37h196d8e1_0
pysocks=1.7.1=py37_1
python=3.7.15=h6244533_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py37haa95532_0
pytz=2022.7=py37haa95532_0
pywavelets=1.3.0=py37h2bbff1b_0
pywin32=305=py37h2bbff1b_0
pywinpty=2.0.2=py37h5da7b33_0
pyyaml=6.0=py37h2bbff1b_1
pyzmq=23.2.0=py37hd77b12b_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_4
qtconsole=5.3.2=py37haa95532_0
qtpy=2.2.0=py37haa95532_0
qtwebkit=5.212=h3ad3cdb_4
requests=2.28.1=py37haa95532_0
scikit-image=0.19.3=py37hd77b12b_1
scikit-learn=1.0.2=py37hf11a4ad_1
scipy=1.7.3=py37h7a0a035_2
seaborn=0.12.2=py37haa95532_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=65.5.0=py37haa95532_0
simpleitk=2.2.1=pypi_0
sip=6.6.2=py37hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h6c2663c_0
sniffio=1.2.0=py37haa95532_1
soupsieve=2.3.2.post1=py37haa95532_0
sqlite=3.40.0=h2bbff1b_0
statsmodels=0.13.5=py37h080aedc_0
tbb=2021.6.0=h59b6b97_0
terminado=0.17.1=py37haa95532_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.7.2=pyhd3eb1b0_2
tinycss2=1.2.1=py37haa95532_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py37haa95532_0
toolz=0.12.0=py37haa95532_0
tornado=6.2=py37h2bbff1b_0
tqdm=4.64.1=py37haa95532_0
traitlets=5.7.1=py37haa95532_0
typing-extensions=4.4.0=py37haa95532_0
typing_extensions=4.4.0=py37haa95532_0
urllib3=1.26.13=py37haa95532_0
utfcpp=3.2.1=haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
vtk=9.0.3=py37h84a6d72_201
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
websocket-client=0.58.0=py37haa95532_4
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py37haa95532_0
win32_setctime=1.1.0=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37haa95532_2
winpty=0.4.3=4
xz=5.2.8=h8cc25b3_0
yaml=0.2.5=he774522_0
zeromq=4.3.4=hd77b12b_0
zfp=0.5.5=hd77b12b_6
zipp=3.8.0=py37haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
